make it better

+++ routes/todos.js
@@ -5,23 +5,27 @@ const Joi = require("joi");
 
 const router = express.Router();
 
-router.get("/", async (req, res) => {
+router.get("/", auth, async (req, res) => {
   try {
     const todos = await Todo.find().sort({ date: -1 });
-    res.send(todos);
+    const filteredTodos = todos.filter((todo) => todo.uid == req.user._id);
+    res.send(filteredTodos);
   } catch (error) {
     res.status(500).send(error);
     console.log(error);
   }
 });
 
-router.delete("/:id", async (req, res) => {
+router.delete("/:id", auth, async (req, res) => {
   // deleteOne()
   // deleteMany()
   // findByIDAndDelete()
   try {
     const todo = await Todo.findById(req.params.id);
     if (!todo) return res.status(404).send("Todo not found");
+
+    if (req.user._id != todo.uid)
+      return res.status(401).send("Todo delete failed. Not authorized");
 
     const deletedTodo = await Todo.findByIdAndDelete(req.params.id);
     res.send(deletedTodo);
@@ -31,7 +35,7 @@ router.delete("/:id", async (req, res) => {
   }
 });
 
-router.post("/", async (req, res) => {
+router.post("/", auth, async (req, res) => {
   const schema = Joi.object({
     name: Joi.string().min(3).max(200).required(),
     author: Joi.string().min(3).max(30),
@@ -68,7 +72,7 @@ router.post("/", async (req, res) => {
   }
 });
 
-router.put("/:id", async (req, res) => {
+router.put("/:id", auth, async (req, res) => {
   const schema = Joi.object({
     name: Joi.string().min(3).max(200).required(),
     author: Joi.string().min(3).max(30),
@@ -81,8 +85,11 @@ router.put("/:id", async (req, res) => {
   if (error) return res.status(400).send(error.details[0].message);
 
   try {
-    const todo = Todo.findById(req.params.id);
+    const todo = await Todo.findById(req.params.id);
     if (!todo) return res.status(404).send("ToDo not found");
+
+    if (req.user._id !== todo.uid)
+      return res.status(401).send("Todo update failed. Not authorized");
 
     const { name, author, isComplete, date, uid } = req.body;
 
@@ -104,32 +111,13 @@ router.put("/:id", async (req, res) => {
   }
 });
 
-router.patch("/:id", async (req, res) => {
+router.patch("/:id", auth, async (req, res) => {
   try {
     const todo = await Todo.findById(req.params.id);
-
     if (!todo) return res.status(404).send("Todo not found");
 
-    const updatedTodo = await Todo.findByIdAndUpdate(
-      req.params.id,
-      {
-        isComplete: !todo.isComplete,
-      },
-      { new: true }
-    );
-    console.log(updatedTodo);
-    res.send(updatedTodo);
-  } catch (error) {
-    res.status(500).send(error.message);
-    console.log(error.message);
-  }
-});
-
-router.delete("/:id", async (req, res) => {
-  try {
-    const todo = await Todo.findById(req.params.id);
-
-    if (!todo) return res.status(404).send("Todo not found");
+    if (req.user._id != todo.uid)
+      return res.status(401).send("Todo check update failed. Not authorized");
 
     const updatedTodo = await Todo.findByIdAndUpdate(
       req.params.id,
